split http_client into trait and implementation files
